Extract the local trace reader into a shared harness module
The spool parsing, span assembly, and trace listing that lived inside
the `eve traces` CLI command move to #harness/local-trace-reader so
other consumers (the dev TUI, tests) can read the local spool without
going through the CLI. The command keeps its behavior and re-exports
the reader API; traces also expose their session ids and window index.
