+++ .changeset/shared-trace-reader.md
@@ -1,0 +1,5 @@
+---
+"eve": patch
+---
+
+The local trace spool reader behind `eve traces` moved into a shared internal module; command behavior is unchanged.
